backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
openai==1.6.0
python-dotenv==1.0.0
